# Elegent-Interfacee/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UQL Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export type DatabaseType = (typeof DatabaseType)[keyof typeof DatabaseType];

export const DatabaseType = {
  relational: "relational",
  document: "document",
  graph: "graph",
} as const;

export interface Database {
  id: number;
  name: string;
  type: DatabaseType;
  description?: string;
  createdAt: string;
  collectionCount: number;
}

export type CreateDatabaseRequestType =
  (typeof CreateDatabaseRequestType)[keyof typeof CreateDatabaseRequestType];

export const CreateDatabaseRequestType = {
  relational: "relational",
  document: "document",
  graph: "graph",
} as const;

export interface CreateDatabaseRequest {
  name: string;
  type: CreateDatabaseRequestType;
  description?: string;
}

export type CollectionType =
  (typeof CollectionType)[keyof typeof CollectionType];

export const CollectionType = {
  table: "table",
  collection: "collection",
  node: "node",
  edge: "edge",
} as const;

export type CollectionSchema = { [key: string]: unknown };

export interface Collection {
  id: number;
  databaseId: number;
  name: string;
  type: CollectionType;
  recordCount: number;
  schema?: CollectionSchema;
}

export interface ExecuteQueryRequest {
  query: string;
  databaseId?: number;
}

export type QueryResultDataItem = { [key: string]: unknown };

export type QueryResultQueryType =
  (typeof QueryResultQueryType)[keyof typeof QueryResultQueryType];

export const QueryResultQueryType = {
  relational: "relational",
  document: "document",
  graph: "graph",
  unknown: "unknown",
} as const;

export interface QueryResult {
  success: boolean;
  data?: QueryResultDataItem[];
  columns?: string[];
  rowCount?: number;
  executionTimeMs: number;
  error?: string;
  queryType?: QueryResultQueryType;
}

export interface QueryHistoryItem {
  id: number;
  query: string;
  databaseId?: number;
  executedAt: string;
  executionTimeMs: number;
  success: boolean;
  rowCount?: number;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateAnthropicConversationBody {
  title: string;
}

export interface SendAnthropicMessageBody {
  content: string;
  databaseId?: number;
  databaseName?: string;
  schema?: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}

export type ListQueryHistoryParams = {
  limit?: number;
};