# Elegent-Interfacee/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: UQL Studio API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: databases
    description: Database management
  - name: queries
    description: Query execution
  - name: history
    description: Query history
  - name: anthropic
    description: Anthropic AI chat operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /databases:
    get:
      operationId: listDatabases
      tags: [databases]
      summary: List all databases
      responses:
        "200":
          description: List of databases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Database"
    post:
      operationId: createDatabase
      tags: [databases]
      summary: Create a new database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDatabaseRequest"
      responses:
        "201":
          description: Created database
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Database"

  /databases/{id}:
    get:
      operationId: getDatabase
      tags: [databases]
      summary: Get database details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Database details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Database"
    delete:
      operationId: deleteDatabase
      tags: [databases]
      summary: Delete a database
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /databases/{id}/collections:
    get:
      operationId: listCollections
      tags: [databases]
      summary: List collections/tables in a database
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of collections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collection"

  /queries/execute:
    post:
      operationId: executeQuery
      tags: [queries]
      summary: Execute a UQL query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteQueryRequest"
      responses:
        "200":
          description: Query result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResult"

  /queries/history:
    get:
      operationId: listQueryHistory
      tags: [history]
      summary: Get query history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Query history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QueryHistoryItem"

  /anthropic/conversations:
    get:
      operationId: listAnthropicConversations
      tags: [anthropic]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnthropicConversation"
    post:
      operationId: createAnthropicConversation
      tags: [anthropic]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnthropicConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicConversation"
  /anthropic/conversations/{id}:
    get:
      operationId: getAnthropicConversation
      tags: [anthropic]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicError"
    delete:
      operationId: deleteAnthropicConversation
      tags: [anthropic]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicError"
  /anthropic/conversations/{id}/messages:
    get:
      operationId: listAnthropicMessages
      tags: [anthropic]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnthropicMessage"
    post:
      operationId: sendAnthropicMessage
      tags: [anthropic]
      summary: Send a message and receive an AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendAnthropicMessageBody"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    Database:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [relational, document, graph]
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        collectionCount:
          type: integer
      required:
        - id
        - name
        - type
        - createdAt
        - collectionCount

    CreateDatabaseRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [relational, document, graph]
        description:
          type: string
      required:
        - name
        - type

    Collection:
      type: object
      properties:
        id:
          type: integer
        databaseId:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [table, collection, node, edge]
        recordCount:
          type: integer
        schema:
          type: object
          additionalProperties: true
      required:
        - id
        - databaseId
        - name
        - type
        - recordCount

    ExecuteQueryRequest:
      type: object
      properties:
        query:
          type: string
        databaseId:
          type: integer
      required:
        - query

    QueryResult:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            type: object
            additionalProperties: true
        columns:
          type: array
          items:
            type: string
        rowCount:
          type: integer
        executionTimeMs:
          type: number
        error:
          type: string
        queryType:
          type: string
          enum: [relational, document, graph, unknown]
      required:
        - success
        - executionTimeMs

    QueryHistoryItem:
      type: object
      properties:
        id:
          type: integer
        query:
          type: string
        databaseId:
          type: integer
        executedAt:
          type: string
          format: date-time
        executionTimeMs:
          type: number
        success:
          type: boolean
        rowCount:
          type: integer
      required:
        - id
        - query
        - executedAt
        - executionTimeMs
        - success

    AnthropicConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    AnthropicMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateAnthropicConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    SendAnthropicMessageBody:
      type: object
      properties:
        content:
          type: string
        databaseId:
          type: integer
        databaseName:
          type: string
        schema:
          type: string
      required:
        - content
    AnthropicConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AnthropicMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    AnthropicError:
      type: object
      properties:
        error:
          type: string
      required:
        - error