# Elegent-Interfacee/lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UQL Studio API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all databases
 */
export const ListDatabasesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  type: zod.enum(["relational", "document", "graph"]),
  description: zod.string().optional(),
  createdAt: zod.date(),
  collectionCount: zod.number(),
});
export const ListDatabasesResponse = zod.array(ListDatabasesResponseItem);

/**
 * @summary Create a new database
 */
export const CreateDatabaseBody = zod.object({
  name: zod.string(),
  type: zod.enum(["relational", "document", "graph"]),
  description: zod.string().optional(),
});

/**
 * @summary Get database details
 */
export const GetDatabaseParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDatabaseResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  type: zod.enum(["relational", "document", "graph"]),
  description: zod.string().optional(),
  createdAt: zod.date(),
  collectionCount: zod.number(),
});

/**
 * @summary Delete a database
 */
export const DeleteDatabaseParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteDatabaseResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary List collections/tables in a database
 */
export const ListCollectionsParams = zod.object({
  id: zod.coerce.number(),
});

export const ListCollectionsResponseItem = zod.object({
  id: zod.number(),
  databaseId: zod.number(),
  name: zod.string(),
  type: zod.enum(["table", "collection", "node", "edge"]),
  recordCount: zod.number(),
  schema: zod.record(zod.string(), zod.unknown()).optional(),
});
export const ListCollectionsResponse = zod.array(ListCollectionsResponseItem);

/**
 * @summary Execute a UQL query
 */
export const ExecuteQueryBody = zod.object({
  query: zod.string(),
  databaseId: zod.number().optional(),
});

export const ExecuteQueryResponse = zod.object({
  success: zod.boolean(),
  data: zod.array(zod.record(zod.string(), zod.unknown())).optional(),
  columns: zod.array(zod.string()).optional(),
  rowCount: zod.number().optional(),
  executionTimeMs: zod.number(),
  error: zod.string().optional(),
  queryType: zod
    .enum(["relational", "document", "graph", "unknown"])
    .optional(),
});

/**
 * @summary Get query history
 */
export const listQueryHistoryQueryLimitDefault = 20;

export const ListQueryHistoryQueryParams = zod.object({
  limit: zod.coerce.number().default(listQueryHistoryQueryLimitDefault),
});

export const ListQueryHistoryResponseItem = zod.object({
  id: zod.number(),
  query: zod.string(),
  databaseId: zod.number().optional(),
  executedAt: zod.date(),
  executionTimeMs: zod.number(),
  success: zod.boolean(),
  rowCount: zod.number().optional(),
});
export const ListQueryHistoryResponse = zod.array(ListQueryHistoryResponseItem);

/**
 * @summary List all conversations
 */
export const ListAnthropicConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
});
export const ListAnthropicConversationsResponse = zod.array(
  ListAnthropicConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateAnthropicConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetAnthropicConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetAnthropicConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteAnthropicConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListAnthropicMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListAnthropicMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.date(),
});
export const ListAnthropicMessagesResponse = zod.array(
  ListAnthropicMessagesResponseItem,
);

/**
 * @summary Send a message and receive an AI response (SSE stream)
 */
export const SendAnthropicMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendAnthropicMessageBody = zod.object({
  content: zod.string(),
  databaseId: zod.number().optional(),
  databaseName: zod.string().optional(),
  schema: zod.string().optional(),
});