# Elegent-Interfacee/lib/api-zod/src/generated/types/collection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UQL Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { CollectionSchema } from "./collectionSchema";
import type { CollectionType } from "./collectionType";

export interface Collection {
  id: number;
  databaseId: number;
  name: string;
  type: CollectionType;
  recordCount: number;
  schema?: CollectionSchema;
}