# Elegent-Interfacee/lib/api-zod/src/generated/types/createDatabaseRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UQL Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateDatabaseRequestType } from "./createDatabaseRequestType";

export interface CreateDatabaseRequest {
  name: string;
  type: CreateDatabaseRequestType;
  description?: string;
}