# Elegent-Interfacee/lib/api-zod/src/generated/types/queryHistoryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UQL Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface QueryHistoryItem {
  id: number;
  query: string;
  databaseId?: number;
  executedAt: Date;
  executionTimeMs: number;
  success: boolean;
  rowCount?: number;
}